# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kouvee Petshop API Documentation",
        "contact": {
            "name": "Natan Hari P.",
            "email": "natanhp@protonmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/customers/restore/{id}": {
            "get": {
                "tags": [
                    "customers"
                ],
                "description": "Restore the delted customer",
                "operationId": "App\\Http\\Controllers\\CustomersController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of customer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted customer"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/employees/restore/{id}": {
            "get": {
                "tags": [
                    "employees"
                ],
                "description": "Restore the delted employee",
                "operationId": "App\\Http\\Controllers\\EmployeesController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of employee",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted employee"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/petsizes/restore/{id}": {
            "get": {
                "tags": [
                    "pet sizes"
                ],
                "description": "Restore the deleted pet size",
                "operationId": "App\\Http\\Controllers\\PetSizesController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a pet size",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted pet size"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/pettypes/restore/{id}": {
            "get": {
                "tags": [
                    "pet types"
                ],
                "description": "Restore the deleted pet type",
                "operationId": "App\\Http\\Controllers\\PetTypesController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a pet type",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted pet type"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/pets/restore/{id}": {
            "get": {
                "tags": [
                    "pets"
                ],
                "description": "Restore the deleted pet",
                "operationId": "App\\Http\\Controllers\\PetsController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a pet",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted pet"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/products/restore/{id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "description": "Restore the delted product",
                "operationId": "App\\Http\\Controllers\\ProductsController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of product",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted product"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/servicedetails/insert": {
            "post": {
                "tags": [
                    "service details"
                ],
                "summary": "Insert a new service detail",
                "operationId": "App\\Http\\Controllers\\ServiceDetailsController::insert",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "PetTypes_id": {
                                        "description": "The id of the pet type",
                                        "type": "int"
                                    },
                                    "PetSizes_id": {
                                        "description": "The id of the pet size",
                                        "type": "int"
                                    },
                                    "Services_id": {
                                        "description": "The id of the service",
                                        "type": "int"
                                    },
                                    "price": {
                                        "description": "The price of the service",
                                        "type": "int"
                                    },
                                    "createdBy": {
                                        "description": "The id of the owner who creates the service",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/noa/servicedetails/getbyid/{id}": {
            "get": {
                "tags": [
                    "service details"
                ],
                "description": "Get an service by id",
                "operationId": "App\\Http\\Controllers\\ServiceDetailsController::getServiceDetailById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the service detail",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Get a service detail by id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/servicedetails/update": {
            "put": {
                "tags": [
                    "service details"
                ],
                "summary": "Update a service detail",
                "operationId": "App\\Http\\Controllers\\ServiceDetailsController::update",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "The id of the service",
                                        "type": "integer"
                                    },
                                    "PetTypes_id": {
                                        "description": "The id of the pet type",
                                        "type": "int"
                                    },
                                    "PetSizes_id": {
                                        "description": "The id of the pet size",
                                        "type": "int"
                                    },
                                    "Services_id": {
                                        "description": "The id of the service",
                                        "type": "int"
                                    },
                                    "price": {
                                        "description": "The price of the service",
                                        "type": "int"
                                    },
                                    "updatedBy": {
                                        "description": "The foreign key of the owner who updates the service",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/servicedetails/delete/{id}": {
            "delete": {
                "tags": [
                    "service details"
                ],
                "summary": "Deletes a service detail",
                "operationId": "App\\Http\\Controllers\\ServiceDetailsController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service detail id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Service detail not deleted it's because either the deletion failed or service detail to be deleted not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/servicedetails/restore/{id}": {
            "get": {
                "tags": [
                    "service details"
                ],
                "description": "Restore the deleted service detail",
                "operationId": "App\\Http\\Controllers\\ServiceDetailsController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a service detail",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted service detail"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/services/restore/{id}": {
            "get": {
                "tags": [
                    "services"
                ],
                "description": "Restore the deleted service",
                "operationId": "App\\Http\\Controllers\\ServicesController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a service",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restore the deleted service"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/suppliers/restore/{id}": {
            "get": {
                "tags": [
                    "suppliers"
                ],
                "description": "Restore the delted supplier",
                "operationId": "App\\Http\\Controllers\\SuppliersController::restore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of supplier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Restored the deleted supplier"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}